+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - configuration: `/etc/bauh`
     - temp dir: `/tmp/bauh@root` (`/tmp/bauh@$USER` for non root users)
     - autostart: `/etc/xdg/autostart` (only used by the Web gem at the moment)
+    - desktop entries: `/usr/share/applications`
     - custom themes: `/usr/share/bauh/themes`
 
 - UI